+++ packages/react-charting/src/components/HorizontalBarChart/HorizontalBarChart.types.ts
@@ -104,6 +104,11 @@ export interface IHorizontalBarChartStyleProps {
    * @default 15
    */
   barHeight?: number;
+
+  /**
+   * prop to check if benchmark data is provided
+   */
+  showTriangle?: boolean;
 }
 
 export interface IHorizontalBarChartStyles {
@@ -116,6 +121,7 @@ export interface IHorizontalBarChartStyles {
    * Styling for each item in the container
    */
   items: IStyle;
+
   /**
    * Style for the chart.
    */
@@ -145,6 +151,11 @@ export interface IHorizontalBarChartStyles {
    * Style for the chart data text denominator.
    */
   chartDataTextDenominator: IStyle;
+
+  /**
+   * Style for the benchmark container
+   */
+  benchmarkContainer: IStyle;
 
   /**
    * Style for the benchmark triangle