+++ packages/react-components/react-toolbar/src/stories/Toolbar/ToolbarMedium.stories.tsx
@@ -1,0 +1,19 @@
+import * as React from 'react';
+import { FontIncrease24Regular, FontDecrease24Regular, TextFont24Regular } from '@fluentui/react-icons';
+import { Toolbar, ToolbarButton } from '@fluentui/react-toolbar';
+import type { ToolbarProps } from '@fluentui/react-toolbar';
+
+export const Medium = (props: Partial<ToolbarProps>) => (
+  <Toolbar
+    {...props}
+    size="medium"
+    style={{
+      border: '2px solid black',
+      borderRadius: '8px',
+    }}
+  >
+    <ToolbarButton aria-label="Increase Font Size" appearance="primary" icon={<FontIncrease24Regular />} />
+    <ToolbarButton aria-label="Decrease Font Size" icon={<FontDecrease24Regular />} />
+    <ToolbarButton aria-label="Reset Font Size" icon={<TextFont24Regular />} />
+  </Toolbar>
+);