+++ scripts/monorepo/index.js
@@ -5,6 +5,7 @@ module.exports = {
   getAllPackageInfo: require('./getAllPackageInfo'),
   isConvergedPackage: require('./isConvergedPackage'),
   getAffectedPackages: require('./getAffectedPackages'),
+  eslintConstants: require('./eslint-constants'),
   getNthCommit: require('./getNthCommit'),
   ...require('./getDefaultEnvironmentVars'),
   ...require('./get-lerna-aliases'),